% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_unreachable_nodes}
\alias{remove_unreachable_nodes}
\title{Remove all unreachable nodes}
\usage{
remove_unreachable_nodes(graph)
}
\arguments{
\item{graph}{A \link[tidygraph:tidygraph]{tidygraph::tidygraph} object}
}
\value{
Tidygraph with all isolated nodes removed.
}
\description{
Remove all nodes that do not have a degree of at least 1.
}
