% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpp.R
\name{rpp}
\alias{rpp}
\title{Find solution to the rural postman problem}
\usage{
rpp(graph, edgeset)
}
\arguments{
\item{graph}{A \link[tidygraph:tidygraph]{tidygraph::tidygraph} object}

\item{edgeset}{A vector of edge ids that match to the edge attribute \code{name} in \code{graph}}
}
\value{
An ordered vector of edge IDs
}
\description{
Find solution to the rural postman problem
}
